test(cnpj): distribute alphanumeric tests into existing test files
Remove the isolated cnpj.alphanumeric.test.ts and spread the new tests
across the existing test files where each class is already covered:

- cnpj.utils.test.ts (new): tests for charToDigitValue, calcVerifyingDigit
  and isRepeatedChars utilities
- cnpj.formatter.test.ts: alphanumeric applyMask/removeMask and symmetry cases
- cnpj.validator.test.ts: autodetection, format: 'alphanumeric', format: 'numeric'
  and invalid-DV cases for alphanumeric CNPJs
- cnpj.generator.test.ts: generate({ format: 'alphanumeric' }) round-trip suite
- cnpj.test.ts: alphanumeric cases added to parse, tryParse, validate, generate,
  toString/toFormatted and immutability describe blocks
